logout button implementation
