+++ src/config/constant.js
@@ -1,2 +1,1 @@
 export const API_SERVER = "http://localhost:5000/api/";
-// export const API_SERVER = "https://api-server-nodejs.appseed.us/api/";
