+++ src/examples/Navbars/DashboardNavbar/index.js
@@ -16,7 +16,7 @@ Coded by www.creative-tim.com
 import { useState, useEffect } from "react";
 
 // react-router components
-import { useLocation, Link } from "react-router-dom";
+import { useLocation } from "react-router-dom";
 
 // prop-types is a library for typechecking of props.
 import PropTypes from "prop-types";
@@ -30,9 +30,7 @@ import Icon from "@mui/material/Icon";
 
 // Soft UI Dashboard React components
 import SuiBox from "components/SuiBox";
-import SuiTypography from "components/SuiTypography";
 import SuiInput from "components/SuiInput";
-import Cube from "examples/Icons/Cube";
 
 // Soft UI Dashboard React example components
 import Breadcrumbs from "examples/Breadcrumbs";
@@ -48,10 +46,6 @@ import { useSoftUIController } from "context";
 import team2 from "assets/images/team-2.jpg";
 import logoSpotify from "assets/images/small-logos/logo-spotify.svg";
 
-import { useAuth } from "../../../auth-context/auth.context";
-import { useHistory } from "react-router-dom";
-import AuthApi from "../../../api/auth";
-
 function DashboardNavbar({ absolute, light, isMini }) {
   const [navbarType, setNavbarType] = useState();
   const [controller, dispatch] = useSoftUIController();
@@ -59,9 +53,7 @@ function DashboardNavbar({ absolute, light, isMini }) {
   const [openMenu, setOpenMenu] = useState(false);
   const classes = styles({ transparentNavbar, absolute, light, isMini });
   const route = useLocation().pathname.split("/").slice(1);
-  const { setUser } = useAuth();
-  let { user } = useAuth();
-  const history = useHistory();
+
   useEffect(() => {
     // Setting the navbar type
     if (fixedNavbar) {
@@ -96,13 +88,6 @@ function DashboardNavbar({ absolute, light, isMini }) {
     dispatch({ type: "OPEN_CONFIGURATOR", value: !openConfigurator });
   const handleOpenMenu = (event) => setOpenMenu(event.currentTarget);
   const handleCloseMenu = () => setOpenMenu(false);
-
-  const handleLogout = async () => {
-    await AuthApi.Logout(user);
-    await setUser(null);
-    localStorage.removeItem("user");
-    return history.push("/authentication/sign-in");
-  };
 
   // Render the notifications menu
   const renderMenu = () => (
@@ -166,20 +151,6 @@ function DashboardNavbar({ absolute, light, isMini }) {
               color={light ? "white" : "inherit"}
               customClass={classes.navbar_section_desktop}
             >
-              <Link to="#">
-                <IconButton onClick={handleLogout} className={classes.navbar_icon_button}>
-                  <Icon className={light ? "text-white" : "text-dark"}>
-                    <Cube size="12px" />
-                  </Icon>
-                  <SuiTypography
-                    variant="button"
-                    fontWeight="medium"
-                    textColor={light ? "white" : "dark"}
-                  >
-                    Sign Out
-                  </SuiTypography>
-                </IconButton>
-              </Link>
               <IconButton
                 size="small"
                 color="inherit"
